+++ korge/src/commonMain/kotlin/korlibs/korge/view/FpsOverlay.kt
@@ -166,9 +166,9 @@ private class TimeSlidingWindow(val capacity: Int) {
     val min: TimeSpan get() = deque.minOrNull()?.microseconds ?: 1.microseconds
     val max: TimeSpan get() = deque.maxOrNull()?.microseconds ?: 1.microseconds
 
-    val avgFps: Float get() = 1.seconds / avg
-    val minFps: Float get() = 1.seconds / max
-    val maxFps: Float get() = 1.seconds / min
+    val avgFps: Float get() = (1.seconds / avg).toFloat()
+    val minFps: Float get() = (1.seconds / max).toFloat()
+    val maxFps: Float get() = (1.seconds / min).toFloat()
 
     operator fun get(index: Int): TimeSpan = deque[index].microseconds
